Update bst.h
* Added all functions EXCEPT BSTIterator functions (find, remove, etc)
* troubleshooted - It runs
* TODO:
   Seg Fault in copyBinaryTree -- Private method in BST

+++ bst.h
@@ -14,6 +14,10 @@
 #include "bnode.h"   // for BinaryNode
 #include "stack.h"   // for Stack
 
+// forward declaration for BST iterators
+template <class T>
+class BSTIterator;
+
 /***********************************************
  * BINARY SEARCH TREE
  * A binary tree that follows the rule pLeft < root < pRight
@@ -29,18 +33,22 @@ class BST
       {
          clear();
       }
-      BST <T> & operator = (const BST <T> & rhs);
+      BST <T> & operator = (const BST <T> & rhs) throw (const char *);
       
       // make a friend
-      friend void deleteBinaryTree(BinaryNode <T> * &pNode);
+      template <class Y>
+      friend void deleteBinaryTree(BinaryNode <Y> * &pNode);
       
       // container interfaces
       bool empty()   { return root == NULL;    }
-      int size()     { return root->size();     }
+      int size()     { return empty() ? 0 : root->size();     }
       void clear()   
       { 
-         deleteBinaryTree(root); 
-         root = NULL;
+         if (root)
+         {
+            deleteBinaryTree(root); 
+            root = NULL;
+         }
       }
       
       // Binary Search Tree interfaces
@@ -56,14 +64,159 @@ class BST
       
    private:
       BinaryNode <T> * root; 
+      
+      // private methods
+      void copyBinaryTree(BinaryNode <T> * & pNode, const BinaryNode <T> * rhs)
+      {
+         if (rhs == NULL)
+            pNode = NULL;
+         try
+         {
+            pNode = new BinaryNode <T>;
+            pNode->data = rhs->data;
+            copyBinaryTree(pNode->pLeft, rhs->pLeft);
+            copyBinaryTree(pNode->pRight, rhs->pRight);
+         }
+         catch(std::bad_alloc)
+         {
+            throw "ERROR: Unable to allocate a node";
+         }
+      }
 };
+
+/***********************************************
+ * BST :: COPY CONSTRUCTOR
+ * Copy rhs to current tree
+ **********************************************/
+template <class T> 
+BST <T> :: BST(const BST <T> & rhs) throw (const char *)
+{
+   copyBinaryTree(root, rhs.root);
+}
+
+/***********************************************
+ * BST :: OPERATOR =
+ * Assign rhs to current tree
+ **********************************************/
+template <class T>
+ BST <T> & BST <T> :: operator = (const BST <T> & rhs) throw (const char *)
+{
+   // check for !empty tree
+   if (!empty())
+      clear();
+   // IF rhs NULL
+   if (rhs.root == NULL)
+      root = NULL;
+   // copy the tree
+   copyBinaryTree(root, rhs.root);
+   
+   return *this;
+}
+
+/***********************************************
+ * BST :: INSERT
+ * Insert a value into the tree
+ **********************************************/
+template <class T>
+void BST <T> :: insert(const T & t) throw (const char *)
+{
+   // empty tree
+   if (empty())
+   {
+      try
+      {
+         root = new BinaryNode <T>(t);
+      }
+      catch (std::bad_alloc)
+      {
+         throw "ERROR: Unable to allocate a node";
+      }
+   }
+   else
+   {
+      BinaryNode <T> * pNode = NULL;
+      try
+      {
+         BinaryNode <T> * pNode = new BinaryNode <T>(t);
+      }
+      catch (std::bad_alloc)
+      {
+         throw "ERROR: Unable to allocate a node";
+      }
+      
+      BinaryNode <T> * pCurrent = root; 
+      // WHILE place not found
+      bool isFound = false;
+      while (!isFound)
+      {
+         // compare & traverse
+         if (pNode->data > pCurrent->data)
+         {
+            if (pCurrent->pRight == NULL)
+            {
+               pCurrent->addRight(pNode);
+               isFound = true;
+            }
+         }
+         else
+         {
+            if (pCurrent->pLeft == NULL)
+            {
+               pCurrent->addLeft(pNode);
+               isFound = true;
+            }
+         }
+      }
+   }
+}
+
+/***********************************************
+ * BST :: REMOVE
+ * Remove a value from the tree
+ **********************************************/
+template <class T>
+void BST <T> :: remove(BSTIterator <T> & it)
+{
+   
+}
+
+/***********************************************
+ * BST :: FIND
+ * Return an iterator to a value in the tree
+ **********************************************/
+template <class T>
+BSTIterator <T> BST <T> :: find(const T & t)
+{
+   return BSTIterator <T> (NULL);
+}
+
+/***********************************************
+ * BST :: BEGIN
+ * Return an iterator to the start of the tree
+ **********************************************/
+template <class T>
+BSTIterator <T> BST <T> :: begin()
+{
+   return BSTIterator <T> (NULL); 
+}
+
+/***********************************************
+ * BST :: RBEGIN
+ * return a reverse iterator
+ **********************************************/
+template <class T>
+BSTIterator <T> BST <T> :: rbegin()
+{
+   return BSTIterator <T> (NULL);
+}
+
 
 /***********************************************
  * BST ITERATOR
  * A class to iterate through the BST
  **********************************************/
 template <class T>
-BSTIterator
+class BSTIterator
 {
    public:
       BSTIterator(BinaryNode <T> * pNode);
@@ -81,6 +234,7 @@ BSTIterator
          return nodes.top();
       }
       
+      BSTIterator <T> & operator -- ();
       BSTIterator <T> & operator ++ ();
       BSTIterator <T>   operator ++ (int postfix);
       
@@ -88,6 +242,48 @@ BSTIterator
       Stack <BinaryNode <T> *> nodes;
 };
  
+template <class T>
+BSTIterator <T> :: BSTIterator(BinaryNode <T> * pNode)
+{
+   
+}
+
+template <class T>
+BSTIterator <T> :: BSTIterator(Stack <BinaryNode <T> *> nodes)
+{
+   
+}
+
+template <class T>
+BSTIterator <T> & BSTIterator <T> :: operator = (const Stack <BinaryNode <T> *> rhs)
+{
+   
+}
+
+template <class T>
+bool BSTIterator <T> :: operator == (const BSTIterator <T> & rhs) const
+{
+   
+}
+
+template <class T>
+bool BSTIterator <T> :: operator != (const BSTIterator <T> & rhs) const
+{
+   
+}
+
+template <class T>
+BSTIterator <T> & BSTIterator <T> :: operator ++ ()
+{
+   
+}
+
+template <class T>
+BSTIterator <T> BSTIterator <T> :: operator ++ (int postfix)
+{
+   
+}
+      
 /**************************************************
  * BST ITERATOR :: DECREMENT PREFIX
  *     advance by one. Note that this implementation uses
